move user-agent function to main
